move scoring to its own file, make placeholder for weighted scoring
